+++ sdk/communication/AzureCommunicationChat/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 1.0.0
+### Breaking Changes
+- Update message parameters updated to `update(message: String, parameters: UpdateChatMessageRequest)`
+- `EventHandler` renamed to `TrouterEventHandler`
+- Removed `AddChatParticipantsRequest`, `UpdateChatThreadRequest`, `CommunicationIdentifierModel`
+
 ## 1.0.0-beta.12 (2021-06-07)
 ### Breaking Changes
 - Changed the way in which options are instantiated for the following classes: `CreateChatThreadOptions`, `DeleteChatThreadOptions`,  `ListChatThreadsOptions`, `AddChatParticipantsOptions`, `DeleteChatMessageOptions`, `GetChatMessageOptions`, `GetChatThreadPropertiesOptions`, `ListChatMessagesOptions`, `ListChatParticipantsOptions`, `ListChatReadReceiptsOptions`, `RemoveChatParticipantOptions`, `SendChatMessageOptions`, `SendChatReadReceiptOptions`, `SendTypingNotificationOptions`, `UpdateChatMessageOptions`, `UpdateChatThreadPropertiesOptions`.